src/requirements.txt
prometheus-client
pythonping
flask
python3-nmap
pyyaml
